move test.util -> test.cli.util
